[FEATURE] PHP8 Support LRN-31488

+++ Makefile
@@ -71,7 +71,11 @@ clean-dist:
 	rm -rf $(DIST_PREFIX)*/
 
 clean-test:
-	rm -f src/tests/junit.xml
+	test ! -f src/tests/junit.xml || rm -f src/tests/junit.xml
+	test ! -f src/tests/coverage.xml || rm -f src/tests/coverage.xml
+	test ! -d src/tests/coverage || rm -rf src/tests/coverage
+	test ! -f .phpunit.result.cache || rm -f .phpunit.result.cache
 
 clean-vendor:
 	test ! -d vendor || rm -r vendor
+	test ! -f composer.lock || rm -f composer.lock
